Initial commit
There's a lot more to do to properly model what Glimmer VM is doing,
including loops, more correct shouldComponentUpdate, and more correctly
avoiding re-running `if` blocks unless the condition changes.

That said, this is a pretty good starting point for a prototype of the
core ideas in the Glimmer VM reactivity model.