Eliminate custom nested head logic
Previously, blocks had a special-cased head that could get special-cased
head atoms.

This was used, in the test scenarios, for HTML attributes.

I realized that this is just a special-case of the generic "insert
foreign content here" problem, so I implemented a solution
to that problem and updated the HTML scenario to use it for
attributes.

I need to update the CursorAdapter interface to take a range
in flush rather than a cursor, which only makes sense
coincidentally here.